requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
pyodbc>=4.0
matplotlib>=3.6
pyyaml>=6.0
